+++ packages/comment-widget/src/comment-item.ts
@@ -12,7 +12,7 @@ import { LS_UPVOTED_COMMENTS_KEY } from './constant';
 import varStyles from './styles/var';
 import { Ref, createRef, ref } from 'lit/directives/ref.js';
 import { CommentReplies } from './comment-replies';
-import { handleCommentAvatar } from "./user-avatar";
+import { getPolicyInstance } from "./avatar/avatar-policy";
 
 export class CommentItem extends LitElement {
   @consume({ context: baseUrlContext })
@@ -233,6 +233,14 @@ export class CommentItem extends LitElement {
 }
 
 customElements.get('comment-item') || customElements.define('comment-item', CommentItem);
+
+function handleCommentAvatar(comment: CommentVo | undefined): string | undefined {
+  const avatarPolicy = getPolicyInstance();
+  if (avatarPolicy === undefined) {
+    return comment?.owner.avatar;
+  }
+  return avatarPolicy.applyCommentPolicy(comment);
+}
 
 declare global {
   interface HTMLElementTagNameMap {
