+++ packages/comment-widget/src/user-avatar.ts
@@ -1,10 +1,8 @@
-import { CommentVo, ReplyVo } from '@halo-dev/api-client';
 import './icons/icon-loading';
 import { LitElement, css, html } from 'lit';
 import { property, state } from 'lit/decorators.js';
 import baseStyles from './styles/base';
 import varStyles from './styles/var';
-import { getPolicyInstance } from "./avatar/avatar-policy";
 
 export class UserAvatar extends LitElement {
   @property({ type: String })
@@ -131,22 +129,6 @@ export class UserAvatar extends LitElement {
 }
 
 customElements.get('user-avatar') || customElements.define('user-avatar', UserAvatar);
-
-export function handleCommentAvatar(comment: CommentVo | undefined): string | undefined {
-  const avatarPolicy = getPolicyInstance();
-  if (avatarPolicy === undefined) {
-    return comment?.owner.avatar;
-  }
-  return avatarPolicy.applyCommentPolicy(comment);
-}
-
-export function handleReplyAvatar(reply: ReplyVo | undefined): string | undefined {
-  const avatarPolicy = getPolicyInstance();
-  if (avatarPolicy === undefined) {
-    return reply?.owner.avatar;
-  }
-  return avatarPolicy.applyReplyPolicy(reply);
-}
 
 declare global {
   interface HTMLElementTagNameMap {
